testing for _data_io
