--- sktime/datasets/tests/test_data_loaders.py
@@ -1,63 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Test functions for data input and output."""
-
-__author__ = ["Sebastiaan Koel", "Emilia Rose"]
-__all__ = []
-
-import pandas as pd
-import pytest
-
-from sktime.datasets import load_uschange
-
-_CHECKS = {
-    "uschange": {
-        "columns": ["Income", "Production", "Savings", "Unemployment"],
-        "len_y": 187,
-        "len_X": 187,
-        "data_types_X": {
-            "Income": "float64",
-            "Production": "float64",
-            "Savings": "float64",
-            "Unemployment": "float64",
-        },
-        "data_type_y": "float64",
-        "data": load_uschange(),
-    },
-}
-
-
-@pytest.mark.parametrize("dataset", sorted(_CHECKS.keys()))
-def test_data_loaders(dataset):
-    """
-    Assert if datasets are loaded correctly.
-
-    dataset: dictionary with values to assert against should contain:
-        'columns' : list with column names in correct order,
-        'len_y'   : lenght of the y series (int),
-        'len_X'   : lenght of the X series/dataframe (int),
-        'data_types_X' : dictionary with column name keys and dtype as value,
-        'data_type_y'  : dtype if y column (string)
-        'data'    : tuple with y series and X series/dataframe if one is not
-                    applicable fill with None value,
-    """
-    checks = _CHECKS[dataset]
-    y = checks["data"][0]
-    X = checks["data"][1]
-
-    if y is not None:
-        assert isinstance(y, pd.Series)
-        assert len(y) == checks["len_y"]
-        assert y.dtype == checks["data_type_y"]
-
-    if X is not None:
-        if len(checks["data_types_X"]) > 1:
-            assert isinstance(X, pd.DataFrame)
-        else:
-            assert isinstance(X, pd.Series)
-
-        assert X.columns.values.tolist() == checks["columns"]
-
-        for col, dt in checks["data_types_X"].items():
-            assert X[col].dtype == dt
-
-        assert len(X) == checks["len_X"]
